app/requirements.txt
Django>=3.1.7,<3.2.0
Pillow>=8.1.0,<8.2.0
djangorestframework==3.11.0
flake8>=3.8.4,<3.9.0
psycopg2>=2.8.6,<2.9.0
# django-braces==1.14.0
# django-embed-video==1.3.2
# django-allauth==0.42.0
dj-rest-auth==2.1.3